// client-id/src/app/core/interfaces/test-process-status-detail.ts
import { EntrpsAddress } from '../models/test-share-models/entrps-address';

/* eslint-disable @typescript-eslint/naming-convention */
export interface TestProcessSttDetail
{
    athrzSn: number;
    athrzReqstNo: number;
    athrzRceptNo: string;
    ktcBrffcCode: string;
    ktcBrffcCodeName: string;
    athrzProcessSttusCode: string;
    athrzProcessSttusName: string;
    proprod: number;
    insttSn: number;
    insttNm: string;
    processPrarnDe: Date;
    entrpsSn: number;
    entrpsNm: string;
    rprsntvNm: string;
    entrpsSeNm: string;
    bizrno: string;
    bplcAdresSeCode: string;
    bplcZip: string;
    bplcBassAdres: string;
    bplcDetailAdres: string;
    bplcTelno: string;
    bplcFxnum: string;
    fctryAdresSeZip: string;
    fctryZip: string;
    fctryBassAdres: string;
    fctryDetailAdres: string;
    fctryTelno: string;
    fctryFxnum: string;
    vesslNoMarkSeCode: string;
    rqstDt_Date: string;
    rm: string;
    rceptDe: string;
    rceptChargerId: string;
    rceptChargerNm: string;
    rceptChargerDeptNm: string;
    rceptConfmDe: Date;
    rceptConfmerId: string;
    rceptConfmerNm: string;
    rceptConfmerDeptNm: string;
    athrzChargerId: string;
    athrzChargerNm: string;
    athrzChargerDeptNm: string;
    athrzChargerTelno: string;
    lastConfmDe: string;
    athrzRspnberId: string;
    athrzRspnberNm: string;
    athrzRspnberDeptNm: string;
    athrzProcessCn: string;
    atchmnflT_AC003024: any[];
    atchmnflT_AC003025: any[];
    fomConfmNoList: formConfmNo[];
    entrprsBplcAdresList: EntrpsAddress[];
    entrprsFctryAdresList: EntrpsAddress[];
    entrprsMeterAdresList: EntrpsAddress[];
    applcntId: string;
    applcntNm: string;
    applcntTelno: string;
    applcntMbtlnum: string;
    applcntMbtlnum2: string;
    applcntEmailAdres: string;
    issuReqstNbfps: number;
    issuFeeAmount: number;
    issuPrposCn: string;
    outptIndictSeCode: string;
    outptIndictSeCodeName: string;
    subsum: number;
    vat: number;
    totalSmAm: number;
    payAt: string;
    feePayDe: string;
    feePayCnfrmrId: string;
    feePayCnfrmrNm: string;
    feePayCnfrmrDeptNm: string;
    athrzReqstCn: string;
    bsrpDeCtDaycnt: number;
    bsrpNmprCo: number;
    bsrpDeCtAmount: number;
    bsrpStayngCtDaycnt: number;
    bsrpStayngCtAmount: number;
    btrpsSm: number;
    tnAthrzfeeAmount: number;
    postFee: number;
    eqpRntfee: number;
    athrzCnfrmnOutptAmount: number;
    etcamount: number;
}

export interface formConfmNo{
    fomConfmSn: number;
    fomConfmConfmNo: string;
    mrnrSeCode: string;
}